Add new study about safety and efficacy of the AstraZeneca vaccine (or lack thereof)
This partly AstraZeneca-funded study shouldn't be used for approval. I hope it isn't. You don't just take a vaccine that you ADMITTED needed a further trial, which I presume is now taking place, and in the meanwhile just publish a study where you play with the statistics enough to assuage the regulator so that you can start using it anyway before the new trial is over, which is what the European Union has already hinted that it will be doing. (That does make me biased against it due to entirely reasonable factors, and upset that it may be pushed on with anyway, also due to the same reasonable factors.)