Add AstraZeneca "issues" news
I think it's big enough to go into the news section (not adding it to the AZ section itself, it's just a DW article). Picked DW because it lists the countries that are halting use exhaustively, although I'd like a better source that specifies, for example, that a person died in Italy, and another in Denmark, aside from the one in Austria, while the word "death" is only there once in this article.