Adding news of AstraZeneca suspensions in Europe
Including the links to official medicine agencies of countries that have statements about this in English. Please tell me if you know of other ones. I'm now adding this both in the "News" and the AstraZeneca-specific section even though I hadn't added it before as it seems too big now and the German agency quite explicitly suggests a connection with the vaccine, and elaborates on the specific kind of blood clotting events involved.