Attempt to limit amount of low-quality Reddit posts
r/Coronavirus and even more so r/COVID19 quickly downvote things that are highly dubious, but the Rising feed with a limit of 5 doesn't reflect that enough. Switching to Hot with limit 3, although none of these feeds ever seem to actually reflect the sort order I see on the actual website.